Use daemon threads for auto-reload
Use dedicated daemon threads for auto-reload logic so that it doesn't
prevent orderly JVM shutdown.  Also names the thread 'enlive
auto-reloader: $ns'.

Prior to this patch, `clojure.core/future` was used to run the code that
waits for file changes for a ns.  `future` puts tasks on the
agent-send-off thread pool, which consisits of non-daemon threads.
There is no shutdown/stop logic for the auto-reload function, so
auto-reload threads will run forever and prevent orderly shutdown of the
JVM.  Daemon threads don't block JVM shutdown, so this seems like a good
place to use them.